vm-service-updater/requirements.txt
edn-format==0.7.2
gitdb==4.0.5
GitPython==3.1.7
ply==3.11
pyRFC3339==1.1
pytz==2020.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
smmap==3.0.4
